requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.1.2
djangorestframework==3.15.2
sqlparse==0.5.1
